python_src/problem/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
httpx
bcrypt==3.2.0
redis
boto3